Remove redundant '@types/webpack-dev-server' from dependencies
The '@types/webpack-dev-server' was removed from both package.json and pnpm-lock.yaml files. This is in line with the note within the package stating that webpack-dev-server provides its own type definitions, making the '@types/webpack-dev-server' package unnecessary and redundant. This will simplify the project's dependencies and streamline the build process.